refactor: extract token utils
